+++ pkg/scalers/azure_servicebus_scaler.go
@@ -23,6 +23,8 @@ const (
 	none         entityType = 0
 	queue        entityType = 1
 	subscription entityType = 2
+	messageCountMetricName    = "messageCount"
+	defaultTargetMessageCount = 5
 )
 
 var azureServiceBusLog = logf.Log.WithName("azure_servicebus_scaler")
@@ -59,15 +61,15 @@ func NewAzureServiceBusScaler(resolvedEnv, metadata, authParams map[string]strin
 func parseAzureServiceBusMetadata(resolvedEnv, metadata, authParams map[string]string, podIdentity string) (*azureServiceBusMetadata, error) {
 	meta := azureServiceBusMetadata{}
 	meta.entityType = none
-	meta.targetLength = defaultTargetQueueLength
+	meta.targetLength = defaultTargetMessageCount
 
 	// get target metric value
-	if val, ok := metadata[queueLengthMetricName]; ok {
-		queueLength, err := strconv.Atoi(val)
+	if val, ok := metadata[messageCountMetricName]; ok {
+		messageCount, err := strconv.Atoi(val)
 		if err != nil {
-			azureServiceBusLog.Error(err, "Error parsing azure queue metadata", "queueLengthMetricName", queueLengthMetricName)
+			azureServiceBusLog.Error(err, "Error parsing azure queue metadata", "messageCount", messageCountMetricName)
 		} else {
-			meta.targetLength = queueLength
+			meta.targetLength = messageCount
 		}
 	}